Compute prediction interval for a new observation

Depending on the current transformation \(h(y)= \{y, \sqrt{y}, y^{2/3}\}\),

$$V(h(y_0)-h(\mu_0))=V(h(y_0))+V(h(\mu_0))$$

is used to compute a prediction interval. The prediction variance consists of a component due to the variance of having a single observation and a prediction variance.

Usage

algo.farrington.threshold(pred,phi,alpha=0.01,skewness.transform="none",y)

Arguments

pred

A GLM prediction object

phi

Current overdispersion parameter (superfluous?)

alpha

Quantile level in Gaussian based CI, i.e. an \((1-\alpha)\cdot 100\%\) confidence interval is computed.

skewness.transform

Skewness correction, i.e. one of "none", "1/2", or "2/3".

y

Observed number

Value

Vector of length four with lower and upper bounds of an \((1-\alpha)\cdot 100\%\) confidence interval (first two arguments) and corresponding quantile of observation y together with the median of the predictive distribution.