Computation of Quality Values for a Surveillance System Result

Computation of the quality values for a surveillance system output.

Usage

algo.quality(sts, penalty = 20)

Arguments

sts

object of class survRes or sts, which includes the state chain and the computed alarm chain

penalty

the maximal penalty for the lag

Value

an object of class "algoQV", which is a list of quality values:

TP

Number of correct found outbreaks.

FP

Number of false found outbreaks.

TN

Number of correct found non outbreaks.

FN

Number of false found non outbreaks.

sens

True positive rate, meaning TP/(FN + TP).

spec

True negative rate, meaning TN/(TN + FP).

dist

Euclidean distance between (1-spec, sens) to (0,1).

lag

Lag of the outbreak recognizing by the system.

Details

The lag is defined as follows: In the state chain just the beginnings of an outbreak chain (outbreaks directly following each other) are considered. In the alarm chain, the range from the beginning of an outbreak until min(next outbreak beginning, penalty) timepoints is considered. The penalty timepoints were chosen, to provide an upper bound on the penalty for not discovering an outbreak. Now the difference between the first alarm by the system and the defined beginning is denoted “the lag”. Additionally outbreaks found by the system are not punished. At the end, the mean of the lags for every outbreak chain is returned as summary lag.

See also

Examples

# Create a test object
disProgObj <- sim.pointSource(p = 0.99, r = 0.5, length = 200, A = 1,
                              alpha = 1, beta = 0, phi = 0,
                              frequency = 1, state = NULL, K = 1.7)

# Let this object be tested from rki1
survResObj <- algo.rki1(disProgObj, control = list(range = 50:200))

# Compute the list of quality values
quality <- algo.quality(survResObj)
quality # the list is printed in matrix form

.opt <- options(xtable.comment = FALSE)
# Format as an "xtable", which is printed with LaTeX markup (by default)
library("xtable")
xtable(quality)
options(.opt)