Generic animation of spatio-temporal objects

Generic function for animation of R objects.

Usage

animate(object, ...)

Arguments

object

The object to animate.

...

Arguments to be passed to methods, such as graphical parameters or time interval options for the snapshots.

See also