Convert disProg object to sts and vice versa

A small helper function to convert a disProg object to become an object of the S4 class sts and vice versa. In the future the sts should replace the disProg class, but for now this function allows for conversion between the two formats.

Usage

disProg2sts(disProgObj, map=NULL)
   sts2disProg(sts)

Arguments

disProgObj

an object of class "disProg"

map

an optional "SpatialPolygons" object

sts

an object of class "sts" to convert

Value

an object of class "sts" or "disProg", respectively.

See also

Examples

  data(ha)
  print(disProg2sts(ha))
  class(sts2disProg(disProg2sts(ha)))