Update method for "epidataCS"

The update method for the "epidataCS" class may be used to modify the hyperparameters \(\epsilon\) (eps.t) and \(\delta\) (eps.s), the indicator matrix qmatrix determining possible transmission between the event types, the numerical accuracy nCircle2Poly of the polygonal approximation, and the endemic covariates from stgrid (including the time intervals). The update method will also update the auxiliary information contained in an "epidataCS" object accordingly, e.g., the vector of potential sources of each event, the influence regions, or the endemic covariates copied from the new stgrid.

Usage

# S3 method for class 'epidataCS'
update(object, eps.t, eps.s, qmatrix, nCircle2Poly, stgrid, ...)

Arguments

object

an object of class "epidataCS".

eps.t

numeric vector of length 1 or corresponding to the number of events in object$events. The event data column eps.t specifies the maximum temporal influence radius (e.g., length of infectious period, time to culling, etc.) of the events.

eps.s

numeric vector of length 1 or corresponding to the number of events in object$events. The event data column eps.s specifies the maximum spatial influence radius of the events.

qmatrix

square indicator matrix (0/1 or TRUE/FALSE) for possible transmission between the event types.

nCircle2Poly

accuracy (number of edges) of the polygonal approximation of a circle.

stgrid

a new data.frame with endemic covariates, possibly transformed from or adding to the original object$stgrid. The grid must cover the same regions as the original, i.e., levels(object$stgrid$tile) must remain identical. See epidataCS for a detailed description of the required format.

...

unused (argument of the generic).

Value

The updated "epidataCS" object.

Author

Sebastian Meyer

See also

Examples

data("imdepi")

## assume different interaction ranges and simplify polygons
imdepi2 <- update(imdepi, eps.t = 20, eps.s = Inf, nCircle2Poly = 16)
    
(s <- summary(imdepi))
(s2 <- summary(imdepi2))
## The update reduced the number of infectives (along time)
## because the length of the infectious periods is reduced. It also 
## changed the set of potential sources of transmission for each
## event, since the interaction is shorter in time but wider in space
## (eps.s=Inf means interaction over the whole observation region).

## use a time-constant grid
imdepi3 <- update(imdepi, stgrid = subset(imdepi$stgrid, BLOCK == 1, -stop))
(s3 <- summary(imdepi3)) # "1 time block"