Determine the k and h values in a standard normal setting

Given a specification of the average run length in the (a)cceptance and (r)ejected setting determine the k and h values in a standard normal setting.

Usage

find.kh(ARLa = 500, ARLr = 7, sided = "one", method = "BFGS", verbose=FALSE)

Arguments

ARLa

average run length in acceptance setting, aka. in control state. Specifies the number of observations before false alarm.

ARLr

average run length in rejection state, aka. out of control state. Specifies the number of observations before an increase is detected (i.e. detection delay)

sided

one-sided cusum scheme

method

Which method to use in the function optim. Standard choice is BFGS, but in some situation Nelder-Mead can be advantageous.

verbose

gives extra information about the root finding process

Value

Returns a list with reference value k and decision interval h.

Details

Functions from the spc package are used in a simple univariate root finding problem.

Examples

if (requireNamespace("spc")) {
    find.kh(ARLa=500,ARLr=7,sided="one")
    find.kh(ARLa=500,ARLr=3,sided="one")
}