Find decision interval for given in-control ARL and reference value

Function to find a decision interval h* for given reference value k and desired ARL \(\gamma\) so that the average run length for a Poisson or Binomial CUSUM with in-control parameter \(\theta_0\), reference value k and is approximately \(\gamma\), i.e. \(\Big| \frac{ARL(h^*) -\gamma}{\gamma} \Big| < \epsilon\), or larger, i.e. \(ARL(h^*) > \gamma \).

Usage

findH(ARL0, theta0, s = 1, rel.tol = 0.03, roundK = TRUE,
      distr = c("poisson", "binomial"), digits = 1, FIR = FALSE, ...)

hValues(theta0, ARL0, rel.tol=0.02, s = 1, roundK = TRUE, digits = 1,
        distr = c("poisson", "binomial"), FIR = FALSE, ...)

Arguments

ARL0

desired in-control ARL \(\gamma\)

theta0

in-control parameter \(\theta_0\)

s

change to detect, see details

distr

"poisson" or "binomial"

rel.tol

relative tolerance, i.e. the search for h* is stopped if \(\Big| \frac{ARL(h^*) -\gamma}{\gamma} \Big| < \) rel.tol

digits

the reference value k and the decision interval h are rounded to digits decimal places

roundK

passed to findK

FIR

if TRUE, the decision interval that leads to the desired ARL for a FIR CUSUM with head start \(\frac{\code{h}}{2}\) is returned

...

further arguments for the distribution function, i.e. number of trials n for binomial cdf

Value

findH returns a vector and hValues returns a matrix with elements

theta0

in-control parameter

h

decision interval

k

reference value

ARL

ARL for a CUSUM with parameters k and h

rel.tol

corresponds to \(\Big| \frac{ARL(h) -\gamma}{\gamma} \Big|\)

Details

The out-of-control parameter used to determine the reference value k is specified as: $$\theta_1 = \lambda_0 + s \sqrt{\lambda_0} $$ for a Poisson variate \(X \sim Po(\lambda)\)

$$\theta_1 = \frac{s \pi_0}{1+(s-1) \pi_0} $$ for a Binomial variate \(X \sim Bin(n, \pi) \)