Find Reference Value

Calculates the reference value k for a Poisson or binomial CUSUM designed to detect a shift from \(\theta_0\) to \(\theta_1\)

Usage

findK(theta0, theta1, distr = c("poisson", "binomial"),
      roundK = FALSE, digits = 1, ...)

Arguments

theta0

in-control parameter

theta1

out-of-control parameter

distr

"poisson" or "binomial"

digits

the reference value k is rounded to digits decimal places

roundK

For discrete data and rational reference value there is only a limited set of possible values that the CUSUM can take (and therefore there is also only a limited set of ARLs). If roundK=TRUE, integer multiples of 0.5 are avoided when rounding the reference value k, i.e. the CUSUM can take more values.

...

further arguments for the distribution function, i.e. number of trials n for the binomial CDF.

Value

Returns reference value k.