Hepatitis A in Berlin

Number of Hepatitis A cases among adult male (age>18) in Berlin 2001-2006. An increase is seen during 2006

data("ha")
data("ha.sts")

Format

ha is a disProg object containing \(290\times 12\) observations starting from week 1 in 2001 to week 30 in 2006. ha.sts is generated from ha by the converter function disProg2sts using a shape file of Berlin (see Examples).

Source

Robert Koch-Institut: SurvStat: https://survstat.rki.de/; Queried on 25 August 2006.

Robert Koch Institut, Epidemiologisches Bulletin 33/2006, p.290.

Examples

## deprecated "disProg" object
data("ha")
ha
plot(aggregate(ha))

## new-style "sts" object
data("ha.sts")
ha.sts
plot(ha.sts, type = observed ~ unit, labels = TRUE)

## conversion of the old "disProg" object 'ha' to the new S4 class "sts"
if (FALSE) {
shpfile <- system.file("shapes/berlin.shp", package="surveillance")
ha.sts <- disProg2sts(ha, map = maptools::readShapePoly(shpfile,IDvar="SNAME"))
## in data("ha.sts"), German umlauts in 'ha.sts@map@data$BEZIRK'
## have been replaced for compatibility
}