HCL-based Heat Colors from the colorspace Package

If package colorspace is available, its heat_hcl function is used to generate a color palette. Otherwise, the similar Heat 2 palette from R's own hcl.colors are used.

This function was exported as hcl.colors in surveillance 1.14.0 - 1.17.0 but is now internal to avoid a name clash with R 3.6.0 (or later), which introduced a function of that name in the base package grDevices.

Usage

.hcl.colors(ncolors = 100, use.color = TRUE)

Arguments

ncolors

the number of colors (>= 1) to be in the palette.

use.color

logical. Should the palette use colors? Otherwise grey levels are returned.

Value

A character vector of ncolors colors.

Examples

barplot(rep(1,10), col = surveillance:::.hcl.colors(10), axes = FALSE)