Generate control Settings for an hhh4 Model

Generate control Settings for an hhh4 Model

Usage

makeControl(f = list(~1), S = list(0, 0, 1), period = 52, offset = 1, ...)

Arguments

f, S, period

arguments for addSeason2formula defining each of the three model formulae in the order (ar, ne, end). Recycled if necessary within mapply.

offset

multiplicative component offsets in the order (ar, ne, end).

...

further elements for the hhh4 control list. The family parameter is set to "NegBin1" by default.

Value

a list for use as the control argument in hhh4.

Examples

makeControl()

## a simplistic model for the fluBYBW data
## (first-order transmission only, no district-specific intercepts)
data("fluBYBW")
mycontrol <- makeControl(
    f = list(~1, ~1, ~t), S = c(1, 1, 3),
    offset = list(population(fluBYBW)),  # recycled -> in all components
    ne = list(normalize = TRUE),
    verbose = TRUE)
str(mycontrol)

if (surveillance.options("allExamples"))
## fit this model
fit <- hhh4(fluBYBW, mycontrol)