Hospitalized Salmonella cases in Germany 2004-2014

Reported number of cases of Salmonella in Germany 2004-2014 (early 2014) that were hospitalized. The corresponding total number of cases is indicated in the slot populationFrac and multinomialTS is TRUE.

Usage

data(salmHospitalized)

Format

An "sts" object.

Source

The data are queried from the Survstat@RKI database of the German Robert Koch Institute (https://survstat.rki.de/).