Generation of Background Noise for Simulated Timeseries

Generation of a cyclic model of a Poisson distribution as background data for a simulated timevector.

The mean of the Poisson distribution is modelled as: $$\mu = \exp(A \sin( frequency \cdot \omega \cdot (t + \phi)) + \alpha + \beta * t + K * state)$$

Usage

sim.seasonalNoise(A = 1, alpha = 1, beta = 0, phi = 0,
                  length, frequency = 1, state = NULL, K = 0)

Arguments

A

amplitude (range of sinus), default = 1.

alpha

parameter to move along the y-axis (negative values not allowed) with alpha > = A, default = 1.

beta

regression coefficient, default = 0.

phi

factor to create seasonal moves (moves the curve along the x-axis), default = 0.

length

number of weeks to model.

frequency

factor to determine the oscillation-frequency, default = 1.

state

if a state chain is entered the outbreaks will be additional weighted by K.

K

additional weight for an outbreak which influences the distribution parameter mu, default = 0.

Value

an object of class seasonNoise which includes the modelled timevector, the parameter mu and all input parameters.

See also

Author

M. Höhle, A. Riebler, C. Lang

Examples

season <- sim.seasonalNoise(length = 300)
plot(season$seasonalBackground,type = "l")

# use a negative timetrend beta
season <- sim.seasonalNoise(beta = -0.003, length = 300)
plot(season$seasonalBackground,type = "l")