Diggle et al (1995) K-function test for space-time clustering

The function stKtest wraps functions in package splancs to perform the K-function based Monte Carlo permutation test for space-time clustering (Diggle et al, 1995) for "epidataCS". The implementation is due to Meyer et al. (2016).

Usage

stKtest(object, eps.s = NULL, eps.t = NULL, B = 199,
        cores = 1, seed = NULL, poly = object$W)

# S3 method for class 'stKtest'
plot(x, which = c("D", "R", "MC"),
     args.D = list(), args.D0 = args.D, args.R = list(), args.MC = list(),
     mfrow = sort(n2mfrow(length(which))), ...)

Arguments

object

an object of class "epidataCS".

eps.s, eps.t

numeric vectors defining the spatial and temporal grids of critical distances over which to evaluate the test. The default (NULL) uses equidistant values from 0 to the smallest eps.s/eps.t value in object$events, but not larger than half the observed spatial/temporal domain.

B

the number of permutations.

cores

the number of parallel processes over which to distribute the requested number of permutations.

seed

argument for set.seed to initialize the random number generator such that results become reproducible (also if cores > 1, see plapply).

poly

the polygonal observation region of the events (as an object handled by xylist). The default object$W might not work since package splancs does not support multi-polygons. In this case, the poly argument can be used to specify a substitute.

x

an "stKtest".

which

a character vector indicating which diagnostic plots to produce. The full set is c("D", "D0", "R", "MC"). The special value which = "stdiagn" means to call the associated splancs function stdiagn.

args.D,args.D0,args.R,args.MC

argument lists for the plot functions persp (for "D" and "D0"), plot.default ("R"), and truehist ("MC"), respectively, to modify the default settings. Ignored if which = "stdiagn".

mfrow

par-setting to layout the plots. Ignored for which = "stdiagn" and if set to NULL.

...

ignored (argument of the generic).

Value

an object of class "stKtest" (inheriting from "htest"), which is a list with the following components:

method

a character string indicating the type of test performed.

data.name

a character string naming the supplied object.

statistic

the sum \(U\) of the standardized residuals \(R(s,t)\).

parameter

the number B of permutations.

p.value

the p-value for the test.

pts

the coordinate matrix of the event locations (for stdiagn.

stK

the estimated K-function as returned by stkhat.

seD

the standard error of the estimated \(D(s,t)\) as returned by stsecal.

mctest

the observed and permutation values of the test statistic as returned by stmctest.

The plot-method invisibly returns NULL.

References

Diggle, P. J.; Chetwynd, A. G.; Häggkvist, R. and Morris, S. E. (1995): Second-order analysis of space-time clustering Statistical Methods in Medical Research, 4, 124-136.

Meyer, S., Warnke, I., Rössler, W. and Held, L. (2016): Model-based testing for space-time interaction using point processes: An application to psychiatric hospital admissions in an urban area. Spatial and Spatio-temporal Epidemiology, 17, 15-25. doi:10.1016/j.sste.2016.03.002 . Eprint: https://arxiv.org/abs/1512.09052.

Author

Sebastian Meyer

See also

Examples

if (requireNamespace("splancs")) {
    data("imdepi")
    imdepiB <- subset(imdepi, type == "B")
    mainpoly <- coordinates(imdepiB$W@polygons[[1]]@Polygons[[5]])
    SGRID <- c(10, 25, 50, 100, 150)
    TGRID <- c(1, 7, 14, 21)
    B <- 19  # limited here for speed
     
    imdBstKtest <- stKtest(imdepiB, eps.s = SGRID, eps.t = TGRID, B = B,
                           cores = 2, seed = 1, poly = list(mainpoly))
    print(imdBstKtest)
    plot(imdBstKtest)
}