Class "stsBP" – a class inheriting from class sts which allows the user to store the results of back-projecting or nowcasting surveillance time series

A class inheriting from class sts, but with additional slots to store the result and associated confidence intervals from back projection of a sts object.

Slots

The slots are as for "sts". However, two additional slots exists.

ci:

An array containing the upper and lower limit of the confidence interval.

lambda:

Back projection component

Methods

The methods are the same as for "sts".

coerce

signature(from = "sts", to = "stsBP"): convert an object of class sts to class stsBP.

Author

M. Höhle