Subsetting "sts" Objects

The [-method extracts parts of an "sts" object using row (time) and column (unit) indices.

Usage

# S4 method for class 'sts'
x[i, j, ..., drop = FALSE]

Arguments

x

an object of class "sts".

i

optional row index (integer or logical vector).

j

optional column index (character, integer, or logical vector).

drop

logical: Should subsetting by j be applied to the map as well? This requires a character index and is disabled by default. It does not affect the other slots: Dimensions are never dropped.

...

ignored.

Value

an object of class "sts".

Details

Row indices are used to select a subset of the original time period. The start and epoch slots of the time series are adjusted accordingly. A warning is issued if an irregular integer sequence is used to extract rows, e.g., x[c(1,2,4),], which could destroy the structure of the time series (freq).

Column indices work as usual when indexing matrices, so may select units by name, position or a vector of booleans. When subsetting columns, population fractions are recomputed if and only if x is no multinomialTS and already contains population fractions.

NA indices are not supported, negative indices are.

Note that a [<- method (i.e., subassignment) is not implemented.

Examples

data("ha.sts")

# Show a (subset of a) single time series
plot(ha.sts[,7])
plot(ha.sts[year(ha.sts)==2006, 7])

# Map a single time point
plot(ha.sts[5*52+26,], type=observed~unit)
plot(ha.sts,           type=observed~unit, tps=5*52+26) # same -> ?stsplot_space

# Restrict the data (and the map) to a subset of the districts
plot(ha.sts[,c("pank","lich")],           type=observed~unit, labels=TRUE)
plot(ha.sts[,c("pank","lich"),drop=TRUE], type=observed~unit, labels=TRUE)