Create an sts object with a given observation date

Function for creating an sts object with a given observation date.

Usage

sts_observation(sts, dateObservation, cut = TRUE)

Arguments

sts

sts-object we want to set at a previous state. Needs to include a reporting triangle.

dateObservation

Date for which we want the state. Needs to be in the reporting triangle dates.

cut

Boolean indicating whether to have 0 counts after the observation date or to simply cut the sts-object

Examples

data("salmAllOnset")
salmAllOnsety2014m01d20 <- sts_observation(salmAllOnset,
  dateObservation="2014-01-20",cut=FALSE)
plot(salmAllOnset)
lines(observed(salmAllOnsety2014m01d20),type="h",col="red")