Xtable quality value object

xtable a single quality value object in a nicely formatted way

Usage

# S3 method for algoQV
xtable(x,caption = NULL, label = NULL, 
      align = NULL, digits = NULL, display = NULL, ...)

Arguments

x

Quality Values object generated with quality

caption

See xtable

label

See xtable

align

See xtable

digits

See xtable

display

See xtable

...

Further arguments (see xtable)

See also

Examples

# Create a test object
disProgObj <- sim.pointSource(p = 0.99, r = 0.5, length = 200, A = 1,
                              alpha = 1, beta = 0, phi = 0,
                              frequency = 1, state = NULL, K = 1.7)

# Let this object be tested from rki1
survResObj <- algo.rki1(disProgObj, control = list(range = 50:200))

# Compute the quality values in a nice formatted way
xtable(algo.quality(survResObj))