Calculation of Average Run Length for discrete CUSUM schemes

Calculates the average run length (ARL) for an upward CUSUM scheme for discrete distributions (i.e. Poisson and binomial) using the Markov chain approach.

Usage

arlCusum(h=10, k=3, theta=2.4, distr=c("poisson","binomial"),
         W=NULL, digits=1, ...)

Arguments

h

decision interval

k

reference value

theta

distribution parameter for the cumulative distribution function (cdf) \(F\), i.e. rate \(\lambda\) for Poisson variates or probability \(p\) for binomial variates

distr

"poisson" or "binomial"

W

Winsorizing value W for a robust CUSUM, to get a nonrobust CUSUM set W > k+h. If NULL, a nonrobust CUSUM is used.

digits

k and h are rounded to digits decimal places

...

further arguments for the distribution function, i.e. number of trials n for binomial cdf

Value

Returns a list with the ARL of the regular (zero-start) and the fast initial response (FIR) CUSUM scheme with reference value k, decision interval h for \(X \sim F(\theta)\), where F is the Poisson or binomial CDF.

ARL

one-sided ARL of the regular (zero-start) CUSUM scheme

FIR.ARL

one-sided ARL of the FIR CUSUM scheme with head start \(\frac{\code{h}}{2}\)

Source

Based on the FORTRAN code of

Hawkins, D. M. (1992). Evaluation of Average Run Lengths of Cumulative Sum Charts for an Arbitrary Data Distribution. Communications in Statistics - Simulation and Computation, 21(4), p. 1001-1020.