Check the residual process of a fitted twinSIR or twinstim

Transform the residual process (cf. the residuals methods for classes "twinSIR" and "twinstim") such that the transformed residuals should be uniformly distributed if the fitted model well describes the true conditional intensity function. Graphically check this using ks.plot.unif. The transformation for the residuals tau is 1 - exp(-diff(c(0,tau))) (cf. Ogata, 1988). Another plot inspects the serial correlation between the transformed residuals (scatterplot between u_i and u_i+1).

checkResidualProcess(object, plot = 1:2, mfrow = c(1,length(plot)), ...)

Arguments

object

an object of class "twinSIR" or "twinstim".

plot

logical (or integer index) vector indicating if (which) plots of the transformed residuals should be produced. The plot index 1 corresponds to a ks.plot.unif to check for deviations of the transformed residuals from the uniform distribution. The plot index 2 corresponds to a scatterplot of \(u_i\) vs. \(u_{i+1}\). By default (plot = 1:2), both plots are produced.

mfrow

see par.

...

further arguments passed to ks.plot.unif.

Value

A list (returned invisibly, if plot = TRUE) with the following components:

tau

the residual process obtained by residuals(object).

U

the transformed residuals which should be distributed as U(0,1).

ks

the result of the ks.test for the uniform distribution of U.

References

Ogata, Y. (1988) Statistical models for earthquake occurrences and residual analysis for point processes. Journal of the American Statistical Association, 83, 9-27

Author

Sebastian Meyer

See also

Examples

data("hagelloch")
fit <- twinSIR(~ household, data = hagelloch)  # a simplistic model
## extract the "residual process", i.e., the fitted cumulative intensities
residuals(fit)
## assess goodness of fit based on these residuals
checkResidualProcess(fit)  # could be better