Creating an object of class disProg

Creates an object of class disProg from a vector with the weeknumber (week) and matrices with the observed number of counts (observed) and the respective state chains (state), where each column represents an individual time series. The matrices neighbourhood and populationFrac provide information about neighbouring units and population proportions.

create.disProg(week, observed, state, start=c(2001,1), freq=52, 
               neighbourhood=NULL, populationFrac=NULL, epochAsDate=FALSE)

Arguments

week

index in the matrix of observations, typically weeks

observed

matrix with parallel time series of counts where rows are time points and columns are the individual time series for unit/area \(i, i=1,\ldots,m\)

state

matrix with corresponding states

start

vector of length two denoting the year and the sample number (week, month, etc.) of the first observation

freq

sampling frequency per year, i.e. 52 for weekly data, 12 for monthly data, 13 if 52 weeks are aggregated into 4 week blocks.

neighbourhood

neighbourhood matrix \(N\) of dimension \(m \times m\) with elements \(n_{ij}=1\) if units \(i\) and \(j\) are adjacent and 0 otherwise

populationFrac

matrix with corresponding population proportions

epochAsDate

interpret the integers in week as Dates. Default is FALSE

Value

object of class disProg

Author

M. Paul

Examples

# create an univariate disProg object 
# read in salmonella.agona data
salmonella <- read.table(system.file("extdata/salmonella.agona.txt", 
                                     package = "surveillance"), header = TRUE)
# look at data.frame
str(salmonella)

salmonellaDisProg <- create.disProg(week = 1:nrow(salmonella), 
                                    observed = salmonella$observed,
                                    state = salmonella$state,
                                    start = c(1990, 1))

# look at disProg object
salmonellaDisProg