Update method for "epidataCS"

The update method for the "epidataCS" class may be used to modify the hyperparameters \(\epsilon\) (eps.t) and \(\delta\) (eps.s), the indicator matrix qmatrix of possible ways of transmission between the event types, and the numerical accuracy nCircle2Poly of the polygonal representation of a circle. The update method will also update the auxiliary information contained in an "epidataCS" object accordingly, e.g., the vector of potential sources of each event, or the polygonal representation of the influence region.

# S3 method for epidataCS
update(object, eps.t, eps.s, qmatrix, nCircle2Poly, ...)

Arguments

object

an object of class "epidataCS".

eps.t

numeric vector of length the number of events in object$events. The event data column eps.t specifies the maximum temporal influence radius (e.g., length of infectious period, time to culling, etc.) of the events.

eps.s

numeric vector of length the number of events in object$events. The event data column eps.s specifies the maximum spatial influence radius of the events.

qmatrix

square indicator matrix (0/1 or TRUE/FALSE) for possible transmission between the event types.

nCircle2Poly

accuracy (number of edges) of the polygonal approximation of a circle.

...

unused (argument of the generic).

Value

The updated "epidataCS" object.

Author

Sebastian Meyer

See also

Examples

data("imdepi")

## assume different interaction ranges and simplify polygons
imdepi2 <- update(imdepi, eps.t = 20, eps.s = Inf, nCircle2Poly = 16)
    
(s <- summary(imdepi))
(s2 <- summary(imdepi2))
## The update reduced the number of infectives (along time)
## because the length of the infectious periods is reduced. It also 
## changed the set of potential sources of transmission for each
## event, since the interaction is shorter in time but wider in space
## (eps.s=Inf means interaction over the whole observation region).