Specify Formulae in a Random Effects HHH Model

The special functions fe and ri are used to specify unit-specific effects of covariates and random intercept terms, respectively, in the component formulae of hhh4.

fe(x, unitSpecific = FALSE, which = NULL, initial = NULL)

ri(type = c("iid","car"), corr = c("none", "all"),
   initial.fe = 0, initial.var = -.5, initial.re = NULL)

Arguments

x

an expression like sin(2*pi*t/52) involving the time variable t, or just 1 for an intercept. In general this covariate expression might use any variables contained in the control$data argument of the parent hhh4 call.

unitSpecific

logical indicating if the effect of x should be unit-specific. This is a convenient shortcut for which = rep(TRUE, nUnits), where nUnits is the number of units (i.e., columns of the "sts" object).

which

vector of logicals indicating which unit(s) should get an unit-specific parameter. For units with a FALSE value, the effect term for x will be zero in the log-linear predictor. Note especially that setting a FALSE value for the intercept term of a unit, e.g., ar = list(f = ~-1 + fe(1, which=c(TRUE, FALSE))) in a bivariate hhh4 model, does not mean that the (autoregressive) model component is omitted for this unit, but that \(\log(\lambda_1) = \alpha_1\) and \(\log(\lambda_2) = 0\), which is usually not of interest. ATM, omitting an autoregressive effect for a specific unit is not possible.
If which=NULL, the parameter is assumed to be the same for all units.

initial

initial values (on internal scale!) for the fixed effects used for optimization. The default (NULL) means to use zeroes.

type

random intercepts either follow an IID or a CAR model.

corr

whether random effects in different components (such as ar and end) should be correlated or not.

initial.fe

initial value for the random intercept mean.

initial.var

initial values (on internal scale!) for the variance components used for optimization.

initial.re

initial values (on internal scale!) for the random effects used for optimization. The default NULL are random numbers from a normal distribution with zero mean and variance 0.001.

See also

Note

These special functions are intended for use in component formulae of hhh4 models and are not exported from the package namespace.

If unit-specific fixed or random intercepts are specified, an overall intercept must be excluded (by -1) in the component formula.