Subsetting "sts" Objects

The [-method extracts parts of an "sts" object using row (time) and column (unit) indices.

Usage

# S4 method for sts
[(x, i, j, ..., drop)

Arguments

x

an object of class "sts".

i

row index (integer or logical vector).

j

column index (character, integer, or logical vector).

...,drop

unused (arguments of the generic).
Dimensions are never dropped.

Value

an object of class "sts".

Details

Row indices are used to select a subset of the original time period. The start and epoch slots of the time series are adjusted accordingly. A warning is issued if an irregular integer sequence is used to extract rows, e.g., x[c(1,2,4),], which could destroy the structure of the time series (freq).

Column indices work as usual when indexing matrices, so may select units by name, position or a vector of booleans. When subsetting columns, population fractions are recomputed if and only if x is no multinomialTS and already contains population fractions.

NA indices are not supported, negative indices are.

Note that a [<- method (i.e., subassignment) is not implemented.

Examples

data("ha.sts")
haagg <- aggregate(ha.sts, nfreq=13)

plot(haagg[, 3])       # Single series
plot(haagg[1:30, 3])   # Somewhat shorter

#Counts at time 20
plot(haagg[20, ], type = observed ~ unit)