Defunct Functions in Package surveillance

The functions listed here are no longer part of surveillance.

Usage

## Defunct in surveillance 1.20.0
qlomax(p, scale, shape)

## Defunct in surveillance 1.18.0
algo.hhh(disProgObj, control=list(
             lambda=TRUE, neighbours=FALSE, linear=FALSE, nseason=0,
             negbin=c("none", "single", "multiple"),
             proportion=c("none", "single", "multiple"), lag.range=NULL
         ), thetastart=NULL, verbose=TRUE)
algo.hhh.grid(disProgObj, control=list(
                 lambda=TRUE, neighbours=FALSE, linear=FALSE, nseason=0,
                 negbin=c("none", "single", "multiple"),
                 proportion=c("none", "single", "multiple"), lag.range=NULL
              ), thetastartMatrix, maxTime=1800, verbose=FALSE)
create.grid(disProgObj, control, params=list(
                epidemic=c(0.1, 0.9, 5), endemic=c(-0.5, 0.5, 3),
                negbin=c(0.3, 12, 10)
            ))

## Defunct in surveillance 1.17.0
compMatrix.writeTable(compMatrix)
correct53to52(disProgObj, firstweek = 1)
enlargeData(disProgObj, range = 1:156, times = 1)
makePlot(outputpath, data = "k1", method = "rki1",
         name, disease, range = 157:339)
readData(abb, week53to52=TRUE, sysPath=TRUE)
test(data = c("k1", "m5"), range = 157:339)
testSim(p = 0.99, r = 0.01, length = 400, A = 1, alpha = 1,
        beta = 0, phi = 0, frequency = 1, state = NULL, K, 
        range = 200:400)
toFileDisProg(disProgObj, toFile)

Details

A qlomax function is provided in package VGAM.

algo.hhh was an early implementation of the HHH regression framework for multivariate time series of infectious disease counts. An improved and considerably extended implementation is provided by the hhh4 function since 2012.

The trivial function compMatrix.writeTable is no longer used (it did nothing more than generating an xtable of its input).

The ancient test function makePlot is no longer used, nor are functions readData (the datasets are still available via data(m1) etc) and correct53to52 / enlargeData (which both only worked for old "disProg" objects with non-matrix elements). enlargeData is still exemplified in the old vignette("surveillance"). test calls of outbreak detection algorithms applied to the old SurvStat datasets can be found in vignette("surveillance"), and testSim is provided as an example in help("sim.pointSource").

Functions related to the old "disProg" class are no longer needed. The effect of toFileDisProg could still be achieved via write.table of as.data.frame(disProg2sts(disProgObj))[c("epoch", "observed", "state")].

See also