Options of the surveillance Package

Query, set or reset options specific to the surveillance package, similar to what options does for global settings.

Usage

surveillance.options(...)
reset.surveillance.options()

Arguments

...

Either empty, or a sequence of option names (as strings), or a sequence of name=value pairs, or a named list of options. Available options are:

stsTickFactors:

A named vector containing tick sizes for the "sts" x-axis relative to par("tcl"). Each entry contains the size at strptime formatting strings. See the help on stsplot_time1 for details.

"%d"

"%W"

"%V"

"%m"

"%Q"

"%Y"

"%G"

colors:

A named list containing plotting color defaults.

nowSymbol

Color of the "now" symbol in stsNC plots. Default: "springgreen4".

piBars

Color of the prediction interval bars in stsNC plots. Default: "orange".

allExamples:

Logical flag queried before running cumbersome computations in help file examples. For interactive() sessions, this option defaults to TRUE. Otherwise, long examples will only be run if the environment variable _R_SURVEILLANCE_ALL_EXAMPLES_ is set (to any value different from "") when attaching the surveillance package. This is to avoid long computations during (daily) CRAN checks.

Value

reset.surveillance.options reverts all options to their default values and (invisibly) returns these in a list.

For surveillance.options, the following holds:

If no arguments are given, the current values of all package options are returned in a list.

If one option name is given, the current value of this option is returned (not in a list, just the value).

If several option names are given, the current values of these options are returned in a list.

If name=value pairs are given, the named options are set to the given values, and the previous values of these options are returned in a list.

Examples

surveillance.options()