Identify Endemic Components in an Intensity Model

The special function cox marks terms in formulae of the functions twinSIR and simEpidata as endemic components, i.e. variables acting multiplicatively on the baseline infection intensity. An illustrative twinSIR call with two epidemic and two endemic covariates is: twinSIR(~B1 + B2 + cox(vaccination) + cox(size), data=myEpidata).

Technically, this function is implemented as function(x) {x} and defined as “special” in terms.formula.

See also