Plot methods for fitted twinstim's

The fitted conditional intensity function from twinstim may be visualized in at least two ways: iafplot plots the fitted interaction functions (as a function of the distance from the host), and intensityplot.twinstim plots the fitted intensity either aggregated over space (evolution over time) or aggregated over time (spatial surface of the cumulated intensity). The plot method for class "twinstim" is just a wrapper for these two functions.

# S3 method for twinstim
plot(x, which, ...)

Arguments

x

an object of class "twinstim".

which

character. Which characteristic of the conditional intensity should be plotted? Possible values are the ones allowed in the functions iafplot and intensityplot.twinstim, e.g. "siaf", or "epidemic proportion". Partial matching is applied.

...

further arguments passed to iafplot or intensityplot.twinstim.

Value

See the documentation of the respective plot functions, iafplot or intensityplot.twinstim.

Author

Sebastian Meyer

Examples

# see the examples for iafplot() and intensityplot.twinstim()