Multivariate Surveillance through independent univariate algorithms

This function takes an sts object and applies an univariate surveillance algorithm to the time series of each observational unit.

<!-- %This is the main function -->
wrap.algo(sts, algo, control,control.hook=function(k, control)
         return(control),verbose=TRUE,...)

<!-- %Derived functions fixing the control object and the "algo" argument -->
farrington(sts, control=list(range=NULL, b=5, w=3, reweight=TRUE,
         verbose=FALSE, alpha=0.05),...)
bayes(sts, control = list(range = range, b = 0, w = 6,
         actY = TRUE,alpha=0.05),...)
rki(sts, control = list(range = range, b = 2, w = 4,
         actY = FALSE),...)
cusum(sts,  control = list(range=range, k=1.04, h=2.26,
         m=NULL, trans="standard",alpha=NULL),...)
glrpois(sts, control = list(range=range,c.ARL=5, S=1,beta=NULL,
         Mtilde=1, M=-1, change="intercept",theta=NULL),...)
glrnb(sts, control = list(range=range,c.ARL=5, mu0=NULL, alpha=0,
         Mtilde=1, M=-1, change="intercept",
         theta=NULL,dir=c("inc","dec"),
         ret=c("cases","value")),...)
outbreakP(sts, control=list(range = range, k=100,
         ret=c("cases","value"),maxUpperboundCases=1e5),...)

Arguments

sts

Object of class sts

algo

Character string giving the function name of the algorithm to call, e.g. "algo.farrington". Calling is done using do.call.

control

Control object as list. Depends on each algorithm.

control.hook

This is a function for handling multivariate objects. This argument is a function function of integer k and the current control object and which returns the appropriate control object for region k.

verbose

Boolean, if TRUE then textual information about the process is given

...

Additional arguments sent to the algo function.

Value

An sts object with the alarm, upperbound, etc. slots filled with the results of independent and univariate surveillance algorithm.

See also

Author

M. Höhle